unshare03: use sizeof(long long) to fix for 32bit
when executing on 32bit compat layer
"nr_open = sizeof(long) * 8" does not evaluate to a large enough value
for /proc/sys/fs/nr_open.
unshare03.c:38: TBROK: Failed to close FILE '/proc/sys/fs/nr_open': EINVAL (22)
Use (long long) instead.

Link: https://lore.kernel.org/ltp/20250424105536.21645-1-akumar@suse.de/
Fixes: fc8be6ed4 ("unshare03: Simplify EMFILE test by using fixed fd limit (sizeof(long) * 8)")
Suggested-by: Petr Vorel <pvorel@suse.cz>

+++ testcases/kernel/syscalls/unshare/unshare03.c
@@ -27,7 +27,7 @@ static void run(void)
 		.exit_signal = SIGCHLD,
 	};
 
-	int nr_open = sizeof(long) * 8;
+	int nr_open = sizeof(long long) * 8;
 
 	SAFE_DUP2(2, nr_open + 1);
